RemoteTask without applicable host should be skipped
If the host being executed on does not share a role with the task being
executed, the task should be skipped for that host. A message indicating
this will be output if --trace is enabled.

Whereas previously a task would fail if hosts were not defined, the new
behavior is to skip any tasks that do not have applicable hosts,
allowing any combination of ENV['ROLES'] and ENV['HOSTS']. Tasks that
can be run on a given host will be run. Others will be skipped.