Create private methods from Rake::RemoteTask#set
Rake::RemoteTask creates methods on Object because that is the only
feasible way to extend rake.  This means there may be collisions with
method names used by other projects.  I encountered a collision between
rake-remote_task and pry over the command_prefix method.

By making the methods private the chances of collision are reduced but
compatibility in use is maintained.  Rake tasks are created at top-level
so they can call private methods on Object just fine (like Object#gsub
for ruby -n).

There is an incompatibility in tests, though.  You must use #send to see
if your set-created methods are set properly.  I also have a pull
request to vlad which allows it to work with this pull request to
rake-remote_task.